requirements.txt
faiss-cpu==1.7.4
langchain==0.0.147 
PyPDF2==3.0.1 
python-dotenv 
streamlit 
tiktoken
openai